server/requirements.txt
Django==5.0
djangorestframework==3.14.0
django-cors-headers==4.3.1
cloudinary==1.36.0
mysqlclient==2.2.1
python-dotenv==1.0.0
djangorestframework-simplejwt==5.3.1
whitenoise==6.8.2
gunicorn==23.0.0